pinctrl: cherryview: Configure HiZ pins to be input when requested as GPIOs
If the pin is in HiZ mode when it is requested as GPIO its value cannot be
read (it always returns 0). In order to cope with the Linux GPIO subsystem
where we do not have such state at all, turn the pin to be input instead.

Reported-by: Jerome Blin <jerome.blin@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
(cherry picked from commit 2479c7300ed943d62c88cd928ddccb4677c71ad4)

Signed-off-by: Ng Wei Tee <wei.tee.ng@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>